+++ Models/GoldAsset.cs
@@ -4,15 +4,14 @@
     {
         public int Id { get; set; }
         public int UserProfileId { get; set; }
-        public decimal WeightInGrams { get; set; } // Вага золота
+        public decimal WeightInGrams { get; set; }
 
         public GoldAsset(string ticker, string sector, decimal averagePrice, decimal weightInGrams)
-            : base(ticker, sector, averagePrice, (int)weightInGrams)
+            : base(ticker, sector, averagePrice, weightInGrams) 
         {
             WeightInGrams = weightInGrams;
         }
 
-        // Можна додати специфічний метод розрахунку вартості золота
         public decimal CalculateGoldValue()
         {
             return AveragePrice * WeightInGrams;
